+++ lib_arch_xt/src/arch_xt_timer_b.cpp
@@ -147,7 +147,7 @@ void ArchXT_TimerB::ioreg_write_handler(reg_addr_t addr, const ioreg_write_t& da
         bool is_chained = (old_clk_mode == TCB_CLKSEL_CLKTCA_gc);
         bool to_chain = (m_clk_mode == TCB_CLKSEL_CLKTCA_gc);
         if (to_chain != is_chained) {
-            ctlreq_data_t d = { .data = &m_timer, .index = (uint32_t)(to_chain ? 1 : 0) };
+            ctlreq_data_t d = { .data = &m_timer, .index = (to_chain ? 1 : 0) };
             device()->ctlreq(AVR_IOCTL_TIMER('A', '0'), AVR_CTLREQ_TCA_REGISTER_TCB, &d);
         }
